+++ src/routes/payments/+page.svelte
@@ -1,246 +1,46 @@
 <script lang="ts">
-  import Msg from '$lib/components/Msg.svelte'
-  import Section from '$lib/components/Section.svelte'
-  import SectionHeading from '$lib/components/SectionHeading.svelte'
-  import Spinner from '$lib/components/Spinner.svelte'
-  import list from '$lib/icons/list.js'
   import PaymentRow from './PaymentRow.svelte'
-  import { fade, slide } from 'svelte/transition'
   import plus from '$lib/icons/plus.js'
   import { translate } from '$lib/i18n/translations.js'
-  import VirtualList from 'svelte-tiny-virtual-list'
-  import { connections$ } from '$lib/streams.js'
-  import FilterSort from '$lib/components/FilterSort.svelte'
-  import SyncRouteData from '$lib/components/SyncRouteData.svelte'
   import { fetchInvoices, fetchTransactions, fetchUtxos } from '$lib/wallets/index.js'
-  import type { Payment } from '$lib/@types/payments.js'
   import type { Filter, Sorters } from '$lib/@types/common.js'
   import { getFilters, getSorters, getTags } from './filters.js'
-  import { storage } from '$lib/services.js'
-  import { STORAGE_KEYS } from '$lib/constants.js'
-  import { getSortedFilteredItems } from '$lib/db/helpers.js'
-  import { anyFiltersApplied } from '$lib/utils.js'
+  import type { Connection } from '$lib/wallets/interfaces.js'
+  import ItemsList from '$lib/components/ItemsList.svelte'
 
-  let processing = false
-  let items: Payment[]
-  let gettingMoreItems = false
-  let noMoreItems = false
-  let filters: Filter[] = getFilters()
-  let sorters: Sorters = getSorters()
-  let tags: string[] = getTags()
+  const filters: Filter[] = getFilters()
+  const sorters: Sorters = getSorters()
+  const tags: string[] = getTags()
 
-  const loadItems = async () => {
-    processing = true
-
-    items = (await getSortedFilteredItems({
-      filters,
-      tags,
-      sort: sorters.applied,
-      limit: 25,
-      offset: 0,
-      table: 'payments'
-    })) as Payment[]
-
-    processing = false
+  const sync = async (connection: Connection) => {
+    await Promise.all([
+      fetchInvoices(connection),
+      fetchTransactions(connection),
+      fetchUtxos(connection)
+    ])
   }
 
-  const getMoreItems = async () => {
-    if (noMoreItems) return
-
-    gettingMoreItems = true
-
-    const moreItems = (await getSortedFilteredItems({
-      filters,
-      tags,
-      sort: sorters.applied,
-      limit: 25,
-      offset: items.length,
-      lastItem: items[items.length - 1],
-      table: 'payments'
-    })) as Payment[]
-
-    if (moreItems.length) {
-      items = [...items, ...moreItems]
-    } else {
-      noMoreItems = true
-    }
-
-    gettingMoreItems = false
+  const button = {
+    href: '/payments/receive',
+    text: $translate('app.labels.receive'),
+    icon: plus
   }
 
-  $: filtersApplied = anyFiltersApplied(filters)
-
-  const handleFilterSortUpdate = () => {
-    noMoreItems = false
-    loadItems()
-    updateStoredFiltersAndSorter()
-  }
-
-  const updateStoredFiltersAndSorter = () => {
-    try {
-      storage.write(STORAGE_KEYS.filters.payments, JSON.stringify(filters))
-      storage.write(STORAGE_KEYS.sorter.payments, JSON.stringify(sorters.applied))
-    } catch (error) {
-      // can't write to storage
-    }
-  }
-
-  let showFullReceiveButton = false
-
-  // need to adjust this if you change the transaction row height
-  const rowSize = 88
-
-  let previousOffset = 0
-  let direction: 'up' | 'down'
-  let processingScroll = false
-
-  const handleTransactionsScroll = (offset: number, diff: number) => {
-    // scrolled to bottom
-    if (offset === diff) {
-      getMoreItems()
-    }
-
-    if (processingScroll) return
-
-    if (offset + 10 < previousOffset) {
-      processingScroll = true
-      requestAnimationFrame(() => {
-        if (direction === 'up') {
-          showFullReceiveButton = true
-        } else {
-          direction = 'up'
-        }
-        processingScroll = false
-      })
-    } else if (offset > previousOffset) {
-      processingScroll = true
-      requestAnimationFrame(() => {
-        if (direction === 'down') {
-          showFullReceiveButton = false
-        } else {
-          direction = 'down'
-        }
-        processingScroll = false
-      })
-    }
-
-    previousOffset = offset
-  }
-
-  let innerHeight: number
-  let virtualList: VirtualList
-
-  $: maxHeight = innerHeight ? innerHeight - 80 - 56 - 80 : 0
-  $: fullHeight = items ? items.length * rowSize : 0
-  $: listHeight = Math.min(maxHeight, fullHeight)
-
-  $: transactionsContainerScrollable = processing
-    ? false
-    : items.length
-    ? fullHeight > listHeight
-    : false
-
-  $: if (items.length) {
-    setTimeout(() => virtualList && virtualList.recomputeSizes(0), 25)
-  }
-
-  const syncPayments = async () => {
-    await Promise.all(
-      connections$.value.map(connection =>
-        Promise.all([
-          fetchInvoices(connection),
-          fetchTransactions(connection),
-          fetchUtxos(connection)
-        ])
-      )
-    )
-
-    loadItems()
-  }
-
-  loadItems()
+  const noItemsMessage = (filtersApplied: boolean) =>
+    $translate(`app.labels.${filtersApplied ? 'no_payments_filtered' : 'no_payments'}`)
 </script>
 
-<svelte:window bind:innerHeight />
-
-<Section>
-  <div class="w-full flex items-center justify-between">
-    <SectionHeading icon={list} />
-    <div class="flex items-center gap-x-2">
-      <SyncRouteData sync={syncPayments} />
-      <FilterSort bind:filters bind:sorters bind:tags on:updated={handleFilterSortUpdate} />
-    </div>
+<ItemsList
+  {filters}
+  {sorters}
+  {tags}
+  {sync}
+  {button}
+  {noItemsMessage}
+  table="payments"
+  rowSize={88}
+>
+  <div slot="row" let:item>
+    <PaymentRow payment={item} />
   </div>
-
-  <div class="w-full flex">
-    {#if processing}
-      <div in:fade={{ duration: 250 }} class="my-4 w-full flex justify-center">
-        <Spinner />
-      </div>
-    {:else if !items.length}
-      <div class="mt-4 mb-2 w-full">
-        <Msg
-          type="info"
-          closable={false}
-          message={$translate(
-            `app.labels.${filtersApplied ? 'no_payments_filtered' : 'no_payments'}`
-          )}
-        />
-      </div>
-    {:else}
-      <div class="w-full flex flex-col justify-center items-center gap-y-2 rounded mt-2">
-        <VirtualList
-          bind:this={virtualList}
-          on:afterScroll={e => handleTransactionsScroll(e.detail.offset, fullHeight - listHeight)}
-          width="100%"
-          height={listHeight}
-          itemCount={items.length}
-          itemSize={rowSize}
-          getKey={index => items[index].id}
-        >
-          <div slot="item" let:index let:style {style}>
-            {@const item = items[index]}
-            <PaymentRow payment={item} />
-          </div>
-        </VirtualList>
-
-        {#if gettingMoreItems}
-          <div class="absolute bottom-1 text-neutral-500">
-            <Spinner size="1.5rem" />
-          </div>
-        {/if}
-      </div>
-    {/if}
-  </div>
-
-  <div
-    class="bottom-0 right-3 w-full flex justify-end mt-2"
-    class:absolute={transactionsContainerScrollable}
-  >
-    <a
-      href="/payments/receive"
-      class:px-2={transactionsContainerScrollable}
-      class:px-4={!transactionsContainerScrollable || showFullReceiveButton}
-      class="no-underline flex items-center rounded-full bg-neutral-900 border-2 border-neutral-50 py-2 hover:shadow-lg hover:shadow-neutral-50 mt-4 w-min hover:bg-neutral-800 relative"
-      on:mouseenter={() => transactionsContainerScrollable && (showFullReceiveButton = true)}
-      on:mouseleave={() => transactionsContainerScrollable && (showFullReceiveButton = false)}
-    >
-      <div class="absolute top-0 right-0 w-full h-full rounded-full overflow-hidden opacity-70">
-        <img src="/images/shell1.png" class="h-full w-full" alt="texture" />
-      </div>
-
-      <div
-        class="w-6 relative"
-        class:-ml-1={!transactionsContainerScrollable || showFullReceiveButton}
-      >
-        {@html plus}
-      </div>
-
-      {#if !transactionsContainerScrollable || showFullReceiveButton}
-        <div class="ml-1 font-semibold relative" in:slide|local={{ axis: 'x' }}>
-          {$translate('app.labels.receive')}
-        </div>
-      {/if}
-    </a>
-  </div>
-</Section>
+</ItemsList>
