+++ src/routes/payments/Summary.svelte
@@ -52,7 +52,7 @@
 
   <div class="flex items-center gap-x-1" class:justify-center={centered}>
     {#if timestamp}
-      {#await formatDate(timestamp, 'hh:mma') then formatted}
+      {#await formatDate(timestamp, 'do MMM hh:mma') then formatted}
         <div class="text-[0.75em] font-semibold mt-1">{formatted}</div>
       {/await}
     {/if}
